modulos
iniciando separación por módulos a modo de practica.
