+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+    "editor.formatOnPaste": true,
+    "editor.formatOnSave": true,
+    "eslint.format.enable": true
+}
