+++ index.html
@@ -76,7 +76,7 @@
             <a href="#inicio"><img src="images/arrow_upward_black_24dp.svg" alt="arrow-up"></a>
         </div>
     </div>
-    <script src="src/main.js"></script>
+    <script src="src/main.js" type="module"></script>
 </body>
 
 </html>
