+++ src/ui.js
@@ -1,0 +1,4 @@
+import { crearListaPokemones } from "./ux.js";
+export function mostrarListaPokemones(){
+    crearListaPokemones();
+}
